requirements.txt
attr==0.3.2
attrs==21.4.0
beautifulsoup4==4.11.1
Django==3.2
djangorestframework==3.14.0
gensim==4.2.0
nltk==3.7
numpy==1.22.3
pandas==1.4.2
python_dateutil==2.8.2
requests==2.28.1
scikit_learn==1.1.2
snscrape==0.4.3.20220106
